+++ pkg/graph/builder_test.go
@@ -1,17 +1,15 @@
 package graph
 
 import (
-	"context"
 	"strconv"
 	"testing"
 	"time"
 
-	helpers "github.com/mariomac/pipes/pkg/test"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/mariomac/pipes/pkg/graph/stage"
 	"github.com/mariomac/pipes/pkg/node"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 const timeout = time.Second
@@ -30,23 +28,23 @@ func TestOptions_BufferLen(t *testing.T) {
 	}
 	nb := NewBuilder(node.ChannelBufferLen(2))
 	startEnded := make(chan struct{})
-	RegisterStart(nb, func(_ context.Context, cfg startConfig) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(nb, func(cfg startConfig) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			out <- 1
 			out <- 2
 			close(startEnded)
 		}, nil
 	})
-	RegisterTerminal(nb, func(_ context.Context, cfg endConfig) (node.TerminalFunc[int], error) {
+	RegisterTerminal(nb, func(cfg endConfig) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {}, nil
 	})
-	graph, err := nb.Build(context.TODO(), config{
+	graph, err := nb.Build(config{
 		Start:     startConfig{Instance: "1"},
 		End:       endConfig{Instance: "2"},
 		Connector: map[string][]string{"1": {"2"}},
 	})
 	require.NoError(t, err)
-	go graph.Run(context.Background())
+	go graph.Run()
 	select {
 	case <-startEnded:
 		//ok!
@@ -74,15 +72,15 @@ func TestCodecs(t *testing.T) {
 		}
 	})
 	type stCfg struct{}
-	RegisterStart(b, func(_ context.Context, _ stCfg) (node.StartFuncCtx[string], error) {
-		return func(_ context.Context, out chan<- string) {
+	RegisterStart(b, func(_ stCfg) (node.StartFunc[string], error) {
+		return func(out chan<- string) {
 			out <- "1"
 			out <- "2"
 			out <- "3"
 		}, nil
 	})
 	type midCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ midCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ midCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for i := range in {
 				out <- i * 2
@@ -92,7 +90,7 @@ func TestCodecs(t *testing.T) {
 	type termCfg struct{}
 	arr := make([]string, 0, 3)
 	done := make(chan struct{})
-	RegisterTerminal(b, func(_ context.Context, _ termCfg) (node.TerminalFunc[string], error) {
+	RegisterTerminal(b, func(_ termCfg) (node.TerminalFunc[string], error) {
 		return func(in <-chan string) {
 			for i := range in {
 				arr = append(arr, i)
@@ -107,13 +105,13 @@ func TestCodecs(t *testing.T) {
 		Term termCfg `nodeId:"term"`
 		Connector
 	}
-	g, err := b.Build(context.TODO(), cfg{Connector: Connector{
+	g, err := b.Build(cfg{Connector: Connector{
 		"st":  []string{"mid"},
 		"mid": []string{"term"},
 	}})
 	require.NoError(t, err)
 
-	go g.Run(context.Background())
+	go g.Run()
 	select {
 	case <-done:
 		//ok!
@@ -132,8 +130,8 @@ func TestIgnore(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -141,7 +139,7 @@ func TestIgnore(t *testing.T) {
 	})
 
 	type DoublerCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -152,7 +150,7 @@ func TestIgnore(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -168,7 +166,7 @@ func TestIgnore(t *testing.T) {
 		Connector
 	}
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Start:  CounterCfg{From: 1, To: 5},
 		Middle: DoublerCfg{},
 		Term:   MapperCfg{Dst: map1},
@@ -177,7 +175,7 @@ func TestIgnore(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -215,25 +213,25 @@ func TestBuildErrors(t *testing.T) {
 	} {
 		t.Run(tc.expected.Error(), func(t *testing.T) {
 			b := NewBuilder()
-			RegisterStart(b, func(_ context.Context, cfg StartCfg) (node.StartFuncCtx[int], error) {
+			RegisterStart(b, func(cfg StartCfg) (node.StartFunc[int], error) {
 				if cfg.fail {
 					return nil, testError("start")
 				}
-				return func(_ context.Context, _ chan<- int) {}, nil
+				return func(_ chan<- int) {}, nil
 			})
-			RegisterMiddle(b, func(_ context.Context, cfg MiddleCfg) (node.MiddleFunc[int, int], error) {
+			RegisterMiddle(b, func(cfg MiddleCfg) (node.MiddleFunc[int, int], error) {
 				if cfg.fail {
 					return nil, testError("middle")
 				}
 				return func(_ <-chan int, _ chan<- int) {}, nil
 			})
-			RegisterTerminal(b, func(_ context.Context, cfg TermCfg) (node.TerminalFunc[int], error) {
+			RegisterTerminal(b, func(cfg TermCfg) (node.TerminalFunc[int], error) {
 				if cfg.fail {
 					return nil, testError("terminal")
 				}
 				return func(_ <-chan int) {}, nil
 			})
-			_, err := b.Build(context.TODO(), tc.cfg)
+			_, err := b.Build(tc.cfg)
 			require.Error(t, err)
 			var tgt testError
 			require.ErrorAs(t, err, &tgt)
@@ -241,56 +239,4 @@ func TestBuildErrors(t *testing.T) {
 		})
 	}
 
-}
-
-func TestBuildContexts(t *testing.T) {
-	type Count struct{}
-	type Multiply struct{}
-	type Forward struct{}
-	type Cfg struct {
-		Start  Count    `nodeId:"s" sendTo:"m"`
-		Middle Multiply `nodeId:"m" sendTo:"t"`
-		Term   Forward  `nodeId:"t"`
-	}
-	b := NewBuilder(node.ChannelBufferLen(100))
-	// This is not the way that contexts are meant to be used but we will try that they are properly passed
-	RegisterStart(b, func(ctx context.Context, cfg Count) (node.StartFuncCtx[int], error) {
-		cnt := ctx.Value("count").(int)
-		return func(_ context.Context, out chan<- int) {
-			for i := 1; i <= cnt; i++ {
-				out <- i
-			}
-		}, nil
-	})
-	RegisterMiddle(b, func(ctx context.Context, cfg Multiply) (node.MiddleFunc[int, int], error) {
-		mul := ctx.Value("multiply").(int)
-		return func(in <-chan int, out chan<- int) {
-			for n := range in {
-				out <- n * mul
-			}
-		}, nil
-	})
-	RegisterTerminal(b, func(ctx context.Context, cfg Forward) (node.TerminalFunc[int], error) {
-		fwd := ctx.Value("forward").(chan int)
-		return func(in <-chan int) {
-			for n := range in {
-				fwd <- n
-			}
-			close(fwd)
-		}, nil
-	})
-
-	fwd := make(chan int, 10)
-	ctx := context.WithValue(context.WithValue(context.WithValue(
-		context.TODO(), "count", 4), "multiply", 3), "forward", fwd)
-	g, err := b.Build(ctx, Cfg{})
-	// the context for Run doesn't have to be the same as the context for Build
-	g.Run(context.TODO())
-	require.NoError(t, err)
-	assert.Equal(t, 3, helpers.ReadChannel(t, fwd, time.Second))
-	assert.Equal(t, 6, helpers.ReadChannel(t, fwd, time.Second))
-	assert.Equal(t, 9, helpers.ReadChannel(t, fwd, time.Second))
-	assert.Equal(t, 12, helpers.ReadChannel(t, fwd, time.Second))
-	n, ok := <-fwd
-	assert.Falsef(t, ok, "did not expect more messages in the forward channel. Got %v", n)
 }