+++ pkg/graph/graph_test.go
@@ -1,14 +1,14 @@
 package graph
 
 import (
-	"context"
 	"testing"
 	"time"
 
-	"github.com/mariomac/pipes/pkg/graph/stage"
-	"github.com/mariomac/pipes/pkg/node"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/mariomac/pipes/pkg/graph/stage"
+	"github.com/mariomac/pipes/pkg/node"
 )
 
 func TestBasic(t *testing.T) {
@@ -19,8 +19,8 @@ func TestBasic(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -30,7 +30,7 @@ func TestBasic(t *testing.T) {
 	type DoublerCfg struct {
 		stage.Instance
 	}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -42,7 +42,7 @@ func TestBasic(t *testing.T) {
 		stage.Instance
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -57,7 +57,7 @@ func TestBasic(t *testing.T) {
 		Connector
 	}
 	map1, map2 := map[int]struct{}{}, map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: []CounterCfg{
 			{From: 1, To: 5, Instance: "c1"},
 			{From: 6, To: 8, Instance: "c2"}},
@@ -75,7 +75,7 @@ func TestBasic(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -88,91 +88,6 @@ func TestBasic(t *testing.T) {
 	assert.Equal(t, map1, map2)
 }
 
-func TestContext(t *testing.T) {
-	b := NewBuilder()
-
-	type ReceiverCfg struct {
-		stage.Instance
-		Input chan int
-	}
-	RegisterStart(b, func(_ context.Context, cfg ReceiverCfg) (node.StartFuncCtx[int], error) {
-		return func(ctx context.Context, out chan<- int) {
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case i := <-cfg.Input:
-					out <- i
-				}
-			}
-		}, nil
-	})
-
-	type ForwarderCfg struct {
-		stage.Instance
-		Out chan int
-	}
-	allClosed := make(chan struct{})
-	RegisterTerminal(b, func(_ context.Context, cfg ForwarderCfg) (node.TerminalFunc[int], error) {
-		return func(in <-chan int) {
-			for n := range in {
-				cfg.Out <- n
-			}
-			close(allClosed)
-		}, nil
-	})
-	type config struct {
-		Starts []ReceiverCfg
-		Term   ForwarderCfg
-		Connector
-	}
-	cfg := config{
-		Starts: []ReceiverCfg{
-			{Instance: "start1", Input: make(chan int, 10)},
-			{Instance: "start2", Input: make(chan int, 10)},
-		},
-		Term: ForwarderCfg{Instance: "end", Out: make(chan int)},
-		Connector: Connector{
-			"start1": []string{"end"},
-			"start2": []string{"end"},
-		},
-	}
-	g, err := b.Build(context.TODO(), &cfg)
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	go g.Run(ctx)
-
-	// The graph works normally
-	cfg.Starts[0].Input <- 123
-	select {
-	case o := <-cfg.Term.Out:
-		assert.Equal(t, 123, o)
-	case <-time.After(timeout):
-		assert.Fail(t, "timeout while waiting for graph to forward items")
-	}
-	cfg.Starts[1].Input <- 456
-	select {
-	case o := <-cfg.Term.Out:
-		assert.Equal(t, 456, o)
-	case <-time.After(timeout):
-		assert.Fail(t, "timeout while waiting for graph to forward items")
-	}
-
-	// after canceling context, the graph should not forward anything
-	cancel()
-
-	cfg.Starts[0].Input <- 789
-	cfg.Starts[1].Input <- 101
-	select {
-	case o := <-cfg.Term.Out:
-		assert.Failf(t, "graph should have been stopped", "unexpected output of the graph: %v", o)
-	default:
-		// OK!
-	}
-
-}
-
 func TestNodeIdAsTag(t *testing.T) {
 	b := NewBuilder()
 
@@ -180,8 +95,8 @@ func TestNodeIdAsTag(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -189,7 +104,7 @@ func TestNodeIdAsTag(t *testing.T) {
 	})
 
 	type DoublerCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -200,7 +115,7 @@ func TestNodeIdAsTag(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -215,7 +130,7 @@ func TestNodeIdAsTag(t *testing.T) {
 		Connector
 	}
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Middle: DoublerCfg{},
 		Term:   MapperCfg{Dst: map1},
@@ -228,7 +143,7 @@ func TestNodeIdAsTag(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -247,8 +162,8 @@ func TestSendsTo(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -256,7 +171,7 @@ func TestSendsTo(t *testing.T) {
 	})
 
 	type DoublerCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -267,7 +182,7 @@ func TestSendsTo(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -281,7 +196,7 @@ func TestSendsTo(t *testing.T) {
 		Term   MapperCfg  `nodeId:"t"`
 	}
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Middle: DoublerCfg{},
 		Term:   MapperCfg{Dst: map1},
@@ -290,7 +205,7 @@ func TestSendsTo(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -309,19 +224,19 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {}, nil
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {}, nil
 	})
 
 	type DoublerCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {}, nil
 	})
 
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {}, nil
 	})
 
@@ -331,7 +246,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m" sendTo:"t"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err := b.Build(context.TODO(), config1{})
+	_, err := b.Build(config1{})
 	assert.Error(t, err)
 
 	// Should fail because the middle node is missing a sendTo
@@ -340,7 +255,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err = b.Build(context.TODO(), config2{})
+	_, err = b.Build(config2{})
 	assert.Error(t, err)
 
 	// Should fail because the middle node is sending to a start node
@@ -349,7 +264,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m"  sendTo:"s"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err = b.Build(context.TODO(), config3{})
+	_, err = b.Build(config3{})
 	assert.Error(t, err)
 
 	// Should fail because a node cannot send to itself
@@ -358,7 +273,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m"  sendTo:"m"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err = b.Build(context.TODO(), config4{})
+	_, err = b.Build(config4{})
 	assert.Error(t, err)
 
 	// Should fail because a destination node does not exist
@@ -367,7 +282,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m" sendTo:"t"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err = b.Build(context.TODO(), config5{})
+	_, err = b.Build(config5{})
 	assert.Error(t, err)
 
 	// Should fail because the middle node does not have any input
@@ -376,7 +291,7 @@ func TestSendsTo_WrongAnnotations(t *testing.T) {
 		Middle DoublerCfg `nodeId:"m" sendTo:"t"`
 		Term   MapperCfg  `nodeId:"t"`
 	}
-	_, err = b.Build(context.TODO(), config6{})
+	_, err = b.Build(config6{})
 	assert.Error(t, err)
 }
 
@@ -398,15 +313,15 @@ func TestEnabled(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
 		}, nil
 	})
 
-	RegisterMiddle(b, func(_ context.Context, c EnableCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(c EnableCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- c.Add + n*2
@@ -417,7 +332,7 @@ func TestEnabled(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -433,7 +348,7 @@ func TestEnabled(t *testing.T) {
 	}
 
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts:  CounterCfg{From: 1, To: 5},
 		Middle1: EnableCfg{Enable: true, Add: 10},
 		Middle2: EnableCfg{Enable: false, Add: 100},
@@ -443,7 +358,7 @@ func TestEnabled(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -462,15 +377,15 @@ func TestForward_Enabled(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
 		}, nil
 	})
 
-	RegisterMiddle(b, func(_ context.Context, c EnableCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(c EnableCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- c.Add + n*2
@@ -481,7 +396,7 @@ func TestForward_Enabled(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -496,7 +411,7 @@ func TestForward_Enabled(t *testing.T) {
 	}
 
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Middle: EnableCfg{Enable: true, Add: 10},
 		Term:   MapperCfg{Dst: map1},
@@ -505,7 +420,7 @@ func TestForward_Enabled(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -524,15 +439,15 @@ func TestForward_Disabled(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
 		}, nil
 	})
 
-	RegisterMiddle(b, func(_ context.Context, c EnableCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(c EnableCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- c.Add + n*2
@@ -543,7 +458,7 @@ func TestForward_Disabled(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -558,7 +473,7 @@ func TestForward_Disabled(t *testing.T) {
 	}
 
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Middle: EnableCfg{Enable: false, Add: 100},
 		Term:   MapperCfg{Dst: map1},
@@ -567,7 +482,7 @@ func TestForward_Disabled(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -586,8 +501,8 @@ func TestForward_Disabled_Nil(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -598,7 +513,7 @@ func TestForward_Disabled_Nil(t *testing.T) {
 		Add int
 	}
 
-	RegisterMiddle(b, func(_ context.Context, c *NillableCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(c *NillableCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- c.Add + n*2
@@ -609,7 +524,7 @@ func TestForward_Disabled_Nil(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -624,7 +539,7 @@ func TestForward_Disabled_Nil(t *testing.T) {
 	}
 
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Term:   MapperCfg{Dst: map1},
 	})
@@ -632,7 +547,7 @@ func TestForward_Disabled_Nil(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -651,8 +566,8 @@ func TestForward_Disabled_Empty(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -662,7 +577,7 @@ func TestForward_Disabled_Empty(t *testing.T) {
 	// If a slice node is tagged as a single node, we won't treat its elements as single nodes but
 	// everything as a node
 	type SliceCfg []int
-	RegisterMiddle(b, func(_ context.Context, c SliceCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(c SliceCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- c[0] + n*2
@@ -673,7 +588,7 @@ func TestForward_Disabled_Empty(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -688,7 +603,7 @@ func TestForward_Disabled_Empty(t *testing.T) {
 	}
 
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Term:   MapperCfg{Dst: map1},
 	})
@@ -696,7 +611,7 @@ func TestForward_Disabled_Empty(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -716,14 +631,14 @@ func TestMultiNode(t *testing.T) {
 		To   int
 	}
 
-	RegisterMultiStart(b, func(_ context.Context, cfg CounterCfg) ([]node.StartFuncCtx[int], error) {
-		return []node.StartFuncCtx[int]{
-			func(_ context.Context, out chan<- int) {
+	RegisterMultiStart(b, func(cfg CounterCfg) ([]node.StartFunc[int], error) {
+		return []node.StartFunc[int]{
+			func(out chan<- int) {
 				for i := cfg.From; i <= cfg.To; i++ {
 					out <- i
 				}
 			},
-			func(_ context.Context, out chan<- int) {
+			func(out chan<- int) {
 				for i := cfg.From; i <= cfg.To; i++ {
 					out <- i + 100
 				}
@@ -732,7 +647,7 @@ func TestMultiNode(t *testing.T) {
 	})
 
 	type DoublerCfg struct{}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -743,7 +658,7 @@ func TestMultiNode(t *testing.T) {
 	type MapperCfg struct {
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -757,7 +672,7 @@ func TestMultiNode(t *testing.T) {
 		Term   MapperCfg  `nodeId:"t"`
 	}
 	map1 := map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 3},
 		Middle: DoublerCfg{},
 		Term:   MapperCfg{Dst: map1},
@@ -766,7 +681,7 @@ func TestMultiNode(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {
@@ -786,8 +701,8 @@ func TestBasic_CombineAnnotations(t *testing.T) {
 		From int
 		To   int
 	}
-	RegisterStart(b, func(_ context.Context, cfg CounterCfg) (node.StartFuncCtx[int], error) {
-		return func(_ context.Context, out chan<- int) {
+	RegisterStart(b, func(cfg CounterCfg) (node.StartFunc[int], error) {
+		return func(out chan<- int) {
 			for i := cfg.From; i <= cfg.To; i++ {
 				out <- i
 			}
@@ -796,7 +711,7 @@ func TestBasic_CombineAnnotations(t *testing.T) {
 
 	type DoublerCfg struct {
 	}
-	RegisterMiddle(b, func(_ context.Context, _ DoublerCfg) (node.MiddleFunc[int, int], error) {
+	RegisterMiddle(b, func(_ DoublerCfg) (node.MiddleFunc[int, int], error) {
 		return func(in <-chan int, out chan<- int) {
 			for n := range in {
 				out <- n * 2
@@ -808,7 +723,7 @@ func TestBasic_CombineAnnotations(t *testing.T) {
 		stage.Instance
 		Dst map[int]struct{}
 	}
-	RegisterTerminal(b, func(_ context.Context, cfg MapperCfg) (node.TerminalFunc[int], error) {
+	RegisterTerminal(b, func(cfg MapperCfg) (node.TerminalFunc[int], error) {
 		return func(in <-chan int) {
 			for n := range in {
 				cfg.Dst[n] = struct{}{}
@@ -823,7 +738,7 @@ func TestBasic_CombineAnnotations(t *testing.T) {
 		Connector
 	}
 	map1, map2 := map[int]struct{}{}, map[int]struct{}{}
-	g, err := b.Build(context.TODO(), config{
+	g, err := b.Build(config{
 		Starts: CounterCfg{From: 1, To: 5},
 		Middle: DoublerCfg{},
 		Term: []MapperCfg{
@@ -837,7 +752,7 @@ func TestBasic_CombineAnnotations(t *testing.T) {
 
 	done := make(chan struct{})
 	go func() {
-		g.Run(context.Background())
+		g.Run()
 		close(done)
 	}()
 	select {